skills/s7cret/binance-enhanced/requirements.txt
# Основные зависимости
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Безопасность
pycryptodome>=3.20.0

# Производительность
aiohttp>=3.9.0
orjson>=3.9.0
jq>=1.6.0

# Telegram-бот и веб-интерфейс
flask>=3.0.0
python-telegram-bot>=20.0
openpyxl>=3.1.0

# Торговые стратегии и аналитика
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0  # Для технических индикаторов

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Дополнительные утилиты
colorama>=0.4.0  # Цветной вывод в консоли
tabulate>=0.9.0  # Форматирование таблиц
progress>=1.6    # Индикаторы прогресса

# Веб-дашборд (опционально)
plotly>=5.18.0   # Графики для дашборда
dash>=2.14.0     # Интерактивный дашборд

# Разработка
black>=23.0.0    # Форматирование кода
flake8>=6.0.0    # Линтинг
mypy>=1.5.0      # Проверка типов
pre-commit>=3.5.0 # Pre-commit хуки